HPT-256: Solrizing ordered page lists to improve performance of the IA BookReader for Paged objects
This commit touches a number of files and adds/changes the following:

Added paged model method to override to_solr so a list of child pages can be solrized with the paged document.
Added 'pages' action to paged controller to access Solr for ordered pages instead of using paged.order_pages.
Made the pages action dual purpose to return a list, or single page if sent an index parameter.
Added update_index action to the pages controller so that pages are solrized on 'reorder' and 'create'.
Changed the BookReader JS to cache a list of page URIs upfront instead of making a call for each page.
Fixed a problem with the BookReader where some icon paths didn't have the correct asset base.
Changed bookreader links to open the 2up view in a new tab.
Adjusted tests to point to 'pages' action; added indexing step so that page additions are propely solrized.