[HPT-451] First look at mocking out models to speed tests.
This commit makes one formerly heavy example run in 0.24 seconds by
avoiding ActiveFedora altogether.  Model tests should be testing the
model code, so we don't need to run it again here.  Is this a
direction I should be going?  If so, there is much left to do, but now
I think I know how to proceed.